+++ src/scripts/constants.py
@@ -29,6 +29,7 @@ CHOICES_MAPPING = {
     "cs": "Výběr",
     "da": "Svarmuligheder",
     "de": "Antwortmöglichkeiten",
+    "el": "Επιλογές",
     "en": "Choices",
     "es": "Opciones",
     "et": "Vastusevariandid",