+++ src/scripts/create_global_mmlu.py
@@ -9,12 +9,13 @@
 # ]
 # ///
 
-"""Create the Global-MMLU Ukrainian knowledge dataset and upload it to the HF Hub."""
+"""Create the Global-MMLU knowledge datasets and upload them to the HF Hub."""
 
 from collections import Counter
 
 import pandas as pd
 from constants import (
+    CHOICES_MAPPING,
     MAX_NUM_CHARS_IN_INSTRUCTION,
     MAX_NUM_CHARS_IN_OPTION,
     MAX_REPETITIONS,
@@ -27,63 +28,68 @@ from sklearn.model_selection import train_test_split
 
 
 def main() -> None:
-    """Create the Global-MMLU Ukrainian knowledge dataset."""
+    """Create the Global-MMLU knowledge datasets."""
     # Define the dataset ID
     repo_id = "CohereLabs/Global-MMLU"
 
-    # Load the dataset
-    dataset = load_dataset(path=repo_id, name="uk")
-    assert isinstance(dataset, DatasetDict)
+    languages = ["uk", "el"]
 
-    # Convert the dataset to dataframes
-    val_df = dataset["dev"].to_pandas()
-    test_df = dataset["test"].to_pandas()
-    assert isinstance(val_df, pd.DataFrame)
-    assert isinstance(test_df, pd.DataFrame)
+    for language in languages:
+        # Load the dataset
+        dataset = load_dataset(path=repo_id, name=language)
+        assert isinstance(dataset, DatasetDict)
 
-    # Process the dataframes
-    val_df = process_split(df=val_df)
-    test_df = process_split(df=test_df)
+        # Convert the dataset to dataframes
+        val_df = dataset["dev"].to_pandas()
+        test_df = dataset["test"].to_pandas()
+        assert isinstance(val_df, pd.DataFrame)
+        assert isinstance(test_df, pd.DataFrame)
 
-    # Create validation split from dev
-    val_size = 256
-    val_df = val_df.sample(n=val_size, random_state=4242, replace=False)
-    val_df = val_df.reset_index(drop=True)
+        # Process the dataframes
+        val_df = process_split(df=val_df, language=language)
+        test_df = process_split(df=test_df, language=language)
 
-    # Create train and test splits from their test set
-    train_size = 1024
-    test_size = 2048
+        # Create validation split from dev
+        val_size = 256
+        val_df = val_df.sample(n=val_size, random_state=4242, replace=False)
+        val_df = val_df.reset_index(drop=True)
 
-    # Split their test into our train and test
-    train_df, final_test_df = train_test_split(
-        test_df, train_size=train_size, random_state=4242, stratify=test_df.category
-    )
+        # Create train and test splits from their test set
+        train_size = 1024
+        test_size = 2048
 
-    # Sample our test set
-    final_test_df = final_test_df.sample(n=test_size, random_state=4242, replace=False)
+        # Split their test into our train and test
+        train_df, final_test_df = train_test_split(
+            test_df, train_size=train_size, random_state=4242, stratify=test_df.category
+        )
 
-    # Reset the index
-    train_df = train_df.reset_index(drop=True)
-    final_test_df = final_test_df.reset_index(drop=True)
+        # Sample our test set
+        final_test_df = final_test_df.sample(
+            n=test_size, random_state=4242, replace=False
+        )
 
-    # Collect datasets in a dataset dictionary
-    dataset = DatasetDict(
-        train=Dataset.from_pandas(train_df, split=Split.TRAIN),
-        val=Dataset.from_pandas(val_df, split=Split.VALIDATION),
-        test=Dataset.from_pandas(final_test_df, split=Split.TEST),
-    )
+        # Reset the index
+        train_df = train_df.reset_index(drop=True)
+        final_test_df = final_test_df.reset_index(drop=True)
 
-    # Create dataset ID
-    dataset_id = "EuroEval/global-mmlu-uk-mini"
+        # Collect datasets in a dataset dictionary
+        dataset = DatasetDict(
+            train=Dataset.from_pandas(train_df, split=Split.TRAIN),
+            val=Dataset.from_pandas(val_df, split=Split.VALIDATION),
+            test=Dataset.from_pandas(final_test_df, split=Split.TEST),
+        )
 
-    # Remove the dataset from Hugging Face Hub if it already exists
-    HfApi().delete_repo(dataset_id, repo_type="dataset", missing_ok=True)
+        # Create dataset ID
+        dataset_id = f"EuroEval/global-mmlu-{language}-mini"
 
-    # Push the dataset to the Hugging Face Hub
-    dataset.push_to_hub(dataset_id, private=True)
+        # Remove the dataset from Hugging Face Hub if it already exists
+        HfApi().delete_repo(dataset_id, repo_type="dataset", missing_ok=True)
+
+        # Push the dataset to the Hugging Face Hub
+        dataset.push_to_hub(dataset_id, private=True)
 
 
-def process_split(df: pd.DataFrame) -> pd.DataFrame:
+def process_split(df: pd.DataFrame, language: str) -> pd.DataFrame:
     """Process a split of the dataset.
 
     Args:
@@ -96,7 +102,7 @@ def process_split(df: pd.DataFrame) -> pd.DataFrame:
     df["category"] = df["subject_category"]
     df = filter_by_length(df=df)
     df = filter_repetitive(df=df)
-    df = add_text_column(df=df)
+    df = add_text_column(df=df, language=language)
     df = df[["text", "label", "category"]]
     df = df.drop_duplicates(inplace=False)
     df = df.reset_index(drop=True)
@@ -157,7 +163,7 @@ def filter_repetitive(df: pd.DataFrame) -> pd.DataFrame:
     ]
 
 
-def add_text_column(df: pd.DataFrame) -> pd.DataFrame:
+def add_text_column(df: pd.DataFrame, language: str) -> pd.DataFrame:
     """Make a `text` column with all the options.
 
     Args:
@@ -168,7 +174,7 @@ def add_text_column(df: pd.DataFrame) -> pd.DataFrame:
     """
     df["text"] = [
         row.question.replace("\n", " ").strip() + "\n"
-        "Варіанти:\n"
+        f"{CHOICES_MAPPING[language]}:\n"
         "a. " + row.option_a.replace("\n", " ").strip() + "\n"
         "b. " + row.option_b.replace("\n", " ").strip() + "\n"
         "c. " + row.option_c.replace("\n", " ").strip() + "\n"