+++ src/scripts/create_multi_wiki_qa.py
@@ -45,6 +45,7 @@ LANGUAGES = [
     "lt",
     "sk",
     "uk",
+    "el",
 ]